Fix ncdump clang warnings. Fix invalid encodings for some non-ascii strings in tests.

+++ ncdump/ncdump.c
@@ -296,7 +296,7 @@ fileopen(const char* path, void** memp, size_t* sizep)
     int status = NC_NOERR;
     int fd = -1;
     int oflags = 0;
-    size_t size = 0;
+    off_t size = 0;
     void* mem = NULL;
     off_t red = 0;
     char* pos = NULL;
